gita_scholar_agent/requirements.txt
# Gita Scholar Agent Dependencies

# Database
supabase==2.3.4
psycopg2-binary==2.9.9

# HTTP & API
requests==2.31.0
httpx==0.26.0

# Web Scraping
beautifulsoup4==4.12.3
lxml==5.1.0
selenium==4.17.2

# Text Processing
fuzzywuzzy==0.18.0
python-Levenshtein==0.25.0
unicodedata2==15.1.0

# Data Analysis
pandas==2.2.0
numpy==1.26.3

# Configuration
python-dotenv==1.0.1

# Report Generation
jinja2==3.1.3

# Utilities
tqdm==4.66.1
colorama==0.4.6
